Use a more evenly spaced set of resolutions
Modify the input processing to generate files at resolutions of
1080p, 720p, 480p, and 360p. This set keeps reasonably close to
halving the number of pixels at each step down in size, while sticking
to common video sizes.

Note: If we want to extend this pattern to larger resolutions in the
future, then 1440p, 2160p, 2880p, 4320p would be the next steps up.