Restrict encoders to one thread per process
It turns out that many of the encoders we're using have automatic internal
parallelism. This was not being properly accounted for, meaning that we
were accidentally launching many more threads than intended, potentially
oversubscribing the system.

The number of threads launched, and the number which could usefully
be applied in our use case (a single 1080p image per process, at most)
varies from encoder to encoder. Therefore the impact on runtime varies
by encoder, invalidating our previous results.

For now, the easiest fix is to disable this internal parallelism, so
that we run one single-threaded encode per CPU thread. I plan to
extend this in future to allow an explicit specification of how many
threads each encoder should use.