Delete temporary files as we go
So far we've been accumulating enormous amounts of files into /tmp
(multiple GB per run). But most of these files are only used for a
short amount of time. By deleting them as soon as possible, the
maximum load on /tmp is massively reduced, which helps a lot when
it's on a ramdisk.